Unsubscribe *all* removed nodes
The blocks nesting was wrong, and it would release the license only
for the last node in the list, if multiple ones are removed.
This commit indents the code properly so that all licenses are
automatically released.

Also, the unsubscribe URL must be "https://" for the unsubscribe
call to work, and HTTP basic auth needs to be used. In principle,
the password should be extracted from the AWS secrets manager,
but this solution works well in the interim.